Update post_hybpiper.sh
Removed the command that trimmed the fasta headers to 8 characters. I removed this because when analyzes include individuals from several runs, it can occur that they have the same I##_T## name, so I added the run suffix, such as: RUN79_I11_T22. Trimming to 8 characters is thus very problematic in the downstream analyses, and it seems that the command removing everything after the "-" character is enough to get rid of the exon name in the header.